+++ tests/type/test_univ.py
@@ -341,6 +341,8 @@ class BitStringTestCase(unittest.TestCase):
 
     def testSet(self):
         assert self.b.clone('Active') == (1,)
+        assert self.b.clone('Urgent') == (0, 1)
+        assert self.b.clone('Urgent, Active') == (1, 1)
         assert self.b.clone("'1010100110001010'B") == (1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0)
         assert self.b.clone("'A98A'H") == (1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0)
         assert self.b.clone(binValue='1010100110001010') == (1, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0)
@@ -349,7 +351,7 @@ class BitStringTestCase(unittest.TestCase):
         assert self.b.clone((1, 0, 1)) == (1, 0, 1)
 
     def testStr(self):
-        assert str(self.b.clone('Urgent,Active')) == '11'
+        assert str(self.b.clone('Urgent')) == '01'
 
     def testRepr(self):
         assert eval(repr(self.b.clone('Urgent,Active')), {'BitString': univ.BitString}) == self.b.clone(