+++ CHANGES.rst
@@ -33,6 +33,7 @@ Revision 0.2.4, released XX-03-2017
   introduces a deviation from original API.
 - Default and optional components en/decoding of Constructed type
   refactored towards better efficiency and more control.
+- Fixed BitString named bits initialization bug.
 
 Revision 0.2.3, released 25-02-2017
 -----------------------------------